support BigInteger

+++ simple-calculator/src/main/java/com/tuanmhoang/calculator/Calculator.java
@@ -1,21 +1,22 @@
 package com.tuanmhoang.calculator;
 
+import java.math.BigInteger;
 import java.util.function.BiFunction;
 
 public class Calculator {
 
-    public static Integer sum(Integer a, Integer b) {
-        BiFunction<Integer, Integer, Integer> func = (x, y) -> x + y;
+    public static BigInteger sum(Integer a, Integer b) {
+        BiFunction<Integer, Integer, BigInteger> func = (x, y) -> BigInteger.valueOf(x + y);
         return func.apply(a, b);
     }
 
-    public static Integer minus(Integer a, Integer b) {
-        BiFunction<Integer, Integer, Integer> func = (x, y) -> x - y;
+    public static BigInteger minus(Integer a, Integer b) {
+        BiFunction<Integer, Integer, BigInteger> func = (x, y) -> BigInteger.valueOf(x - y);
         return func.apply(a, b);
     }
 
-    public static Integer multiply(Integer a, Integer b) {
-        BiFunction<Integer, Integer, Integer> func = (x, y) -> x * y;
+    public static BigInteger multiply(Integer a, Integer b) {
+        BiFunction<Integer, Integer, BigInteger> func = (x, y) -> BigInteger.valueOf(x * y);
         return func.apply(a, b);
     }
